Add files via upload
The 2048 Game is a super fun puzzle game where you combine tiles to reach the number 2048. It's all  about strategy and planning as you slide the tiles in different directions and try to merge them. The
game gets more challenging as the board fills up and you have to think ahead to make the best moves.